chain tests passing
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/branches/filters_dynamic_loading@22575 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: dd5950c35fb20a40b310edc0febfa6e23226a59d

+++ stacks/common/filters/test/test_chain.cpp
@@ -87,12 +87,12 @@ TEST(MultiChannelFilterChain, configuring){
 }
 TEST(FilterChain, configuring){
   double epsilon = 1e-9;
-  filters::FilterChain<double> chain("double");
+  filters::FilterChain<float> chain("float");
   
   EXPECT_TRUE(chain.configure("MeanFilterFloat5"));
  
-  double v1 = 1;
-  double v1a = 9;
+  float v1 = 1;
+  float v1a = 9;
 
   EXPECT_TRUE(chain.update(v1, v1a));
